crewai-price-extractor/requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.30.0
pydantic==2.6.1
python-dotenv==1.0.1

# CrewAI and AI
crewai==0.28.4
langchain==0.1.6
langchain-core==0.1.30
anthropic==0.15.3

# Database
supabase==1.2.1
psycopg2-binary==2.9.9

# Web scraping tools
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.42.0

# Utilities
python-multipart==0.0.6
asyncio==3.4.3 